refactor page builder structure
