+++ packages/ui/src/components/cms/index.ts
@@ -6,5 +6,5 @@ export { default as ProductPageBuilder } from "./ProductPageBuilder";
 export * from "./products";
 export { RangeInput } from "./RangeInput";
 export { default as StyleEditor } from "./StyleEditor";
-export { default as PageBuilder } from "./PageBuilder";
+export { default as PageBuilder } from "./page-builder/PageBuilder";
 export * from "./page-builder";
