+++ packages/ui/src/components/cms/page-builder/index.ts
@@ -11,3 +11,9 @@ export { default as useMediaLibrary } from "./useMediaLibrary";
 export { useArrayEditor } from "./useArrayEditor";
 export { default as CanvasItem } from "./CanvasItem";
 export { default as Palette } from "./Palette";
+export { default as PageBuilder } from "./PageBuilder";
+export { default as PageToolbar } from "./PageToolbar";
+export { default as PageCanvas } from "./PageCanvas";
+export { default as PageSidebar } from "./PageSidebar";
+export { historyStateSchema, reducer } from "./state";
+export { default as usePageBuilderDrag } from "./usePageBuilderDrag";
